requirements.txt
asgiref==3.9.1
attrs==25.3.0
click==8.2.1
dj-database-url==2.3.0
Django==5.2.5
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
uvicorn==0.32.1
whitenoise==6.8.2
